Got all simple portions of game done
Player can now move and game disables player movement if outside of bounds of player area (need to manually add the boxes where you want to show this). Also have detection for when player wins on if they have the food or not. Need to continue by adding UI elements and go onto the next scene if losing scenario or reset scene for winning scenario.